requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
Django==3.0
django-filter==2.4.0
django-mssql-backend==2.8.1
future==0.18.2
html5lib==1.1
pdfrw==0.4
Pillow==8.1.0
pyodbc==4.0.30
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.4
reportlab==3.5.59
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
wfastcgi
PyYAML~=5.4.1
requests~=2.25.1
msal~=1.9.0